Improve data-read from file

+++ src/main/java/com/dev/model/Salesman.java
@@ -2,19 +2,21 @@ package com.dev.model;
 
 import com.dev.model.enumeration.EntityIdentifier;
 
+import java.math.BigDecimal;
+
 public class Salesman extends AbstractPerson {
 
-    private float salary;
+    private BigDecimal salary;
 
     public Salesman() {
         super(EntityIdentifier.SALESMAN);
     }
 
-    public float getSalary() {
+    public BigDecimal getSalary() {
         return salary;
     }
 
-    public void setSalary(float salary) {
+    public void setSalary(BigDecimal salary) {
         this.salary = salary;
     }
 }
